+++ src/amg/HISTORY.rst
@@ -148,7 +148,7 @@ Release History
 
 2.8.0
 ++++++
-* `az grafana folder create`: support parent folder uid argument for nested folder creation
+* `az grafana folder create`: support parent folder argument for nested folder creation
 * `az grafana folder list`: support nested folders
 * `az grafana folder show`: support nested folders
 * `az grafana backup`: include corresponding nested folders when individual folders are specified